Move setup functions to initialization module (ref #137)
A first step to introduce a global strategy about initialization configuration
was to move everything together and make cliquet/__init__.py a lot lighter.
