+++ cliquet/logs.py
@@ -1,76 +1,12 @@
 import os
-from datetime import datetime
 
 import six
 import structlog
-from pyramid.events import NewRequest, NewResponse
 
 from cliquet import utils
 
 
 logger = structlog.get_logger()
-
-
-def setup_logging(config):
-    """Setup structured logging, and emit `request.summary` event on each
-    request, as recommanded by Mozilla Services standard:
-
-    * https://mana.mozilla.org/wiki/display/CLOUDSERVICES/Logging+Standard
-    * http://12factor.net/logs
-    """
-    settings = config.get_settings()
-
-    renderer_klass = config.maybe_dotted(settings['cliquet.logging_renderer'])
-    renderer = renderer_klass(settings)
-
-    structlog.configure(
-        # Share the logger context by thread.
-        context_class=structlog.threadlocal.wrap_dict(dict),
-        # Integrate with Pyramid logging facilities.
-        logger_factory=structlog.stdlib.LoggerFactory(),
-        wrapper_class=structlog.stdlib.BoundLogger,
-        # Setup logger output format.
-        processors=[
-            structlog.stdlib.filter_by_level,
-            structlog.processors.format_exc_info,
-            renderer,
-        ])
-
-    def on_new_request(event):
-        request = event.request
-        # Save the time the request was received by the server.
-        event.request._received_at = utils.msec_time()
-
-        # New logger context, with infos for request summary logger.
-        logger.new(agent=request.headers.get('User-Agent'),
-                   path=event.request.path,
-                   method=request.method,
-                   querystring=dict(request.GET),
-                   uid=request.authenticated_userid,
-                   lang=request.headers.get('Accept-Language'),
-                   auth_type=getattr(request, 'auth_type', None),
-                   errno=None)
-
-    config.add_subscriber(on_new_request, NewRequest)
-
-    def on_new_response(event):
-        response = event.response
-        request = event.request
-
-        # Compute the request processing time in msec (-1 if unknown)
-        current = utils.msec_time()
-        duration = current - getattr(request, '_received_at', current - 1)
-        isotimestamp = datetime.fromtimestamp(current/1000).isoformat()
-
-        # Bind infos for request summary logger.
-        logger.bind(time=isotimestamp,
-                    code=response.status_code,
-                    t=duration)
-
-        # Ouput application request summary.
-        logger.info('request.summary')
-
-    config.add_subscriber(on_new_response, NewResponse)
 
 
 class ClassicLogRenderer(object):
